dap: Extract diagnostics gutter into gutters.rs
